move getUserPrunedDataFromPathname and getPageAndFlowDataFromPathname to service/flow folder
